move show to separate pkg
