+++ cli/azd/internal/cmd/util.go
@@ -38,7 +38,7 @@ func getResourceGroupFollowUp(
 		projectConfig.ResourceGroupName,
 	)
 	if err == nil {
-		suffix := ":\n" + azurePortalLink(portalUrlBase, subscriptionId, resourceGroupName)
+		suffix := ":\n" + AzurePortalLink(portalUrlBase, subscriptionId, resourceGroupName)
 
 		if v, err := strconv.ParseBool(os.Getenv("AZD_DEMO_MODE")); err == nil && v {
 			suffix = "."
@@ -56,7 +56,7 @@ func getResourceGroupFollowUp(
 	return followUp
 }
 
-func azurePortalLink(portalUrlBase, subscriptionId, resourceGroupName string) string {
+func AzurePortalLink(portalUrlBase, subscriptionId, resourceGroupName string) string {
 	if subscriptionId == "" || resourceGroupName == "" {
 		return ""
 	}
